+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.5.0 (in development)
 -----------------------
 - `Properties` instances can now compare equal to `dict`s
+- Gave `Properties` a `copy` method
 
 v0.4.0 (2017-04-22)
 -------------------